Split very slow wasm tests https://bugs.webkit.org/show_bug.cgi?id=320152 rdar://183088481
Reviewed by Tadeu Zagallo and Marcus Plutowski.

JSTests/wasm/stress/type-index-abstract-heap-types.js was very slow and frequently timing out.
Split them into smaller tests.

* JSTests/wasm/stress/type-index-abstract-heap-types-concrete-vs-abstract.js: Added.
(testConcreteVsAbstract):
* JSTests/wasm/stress/type-index-abstract-heap-types-globals-and-tables.js: Added.
(testGlobalsAndTables):
* JSTests/wasm/stress/type-index-abstract-heap-types-nulls-and-casts.js: Added.
(testAbstractNullsAndCasts):
* JSTests/wasm/stress/type-index-abstract-heap-types-subtype-validation.js: Added.
(testSubtypeValidation):
* JSTests/wasm/stress/type-index-abstract-heap-types.js: Removed.

Canonical link: https://commits.webkit.org/317870@main
